fix: updated dark theme
