+++ components/Avatar/Avatar.tsx
@@ -11,7 +11,7 @@ function Avatar() {
         xmlns="http://www.w3.org/2000/svg"
         className={styles.avatar}
       >
-        <path fill="#805BE7" d="M220.677 0h989v989h-989z" />
+        <path d="M220.677 0h989v989h-989z" className={styles.bg} />
         <path
           d="M423.177 649.5c58.8 12.4 112.167-.167 131.5-8 75.5 2.667 232.9 9.6 258.5 16 25.6 6.4 94.333 2.667 125.5 0 70.803-37.6 172.163-34 214.003-27.5 142 63.5 225 267.5 252.5 325.5 22 46.4-48.84 69.33-87 75-460.003 69.33-1362.903 151.4-1294.503-75 68.4-226.4 205.167-303.667 265-314 20.333-2.5 75.7-4.4 134.5 8z"
           fill="#000"
