+++ components/Projects/Project.tsx
@@ -33,12 +33,14 @@ function Project() {
       ))}
       <div className="flex flex-start-top">
         <div>
-          <p className="text-dark">Dive in and learn something new!</p>
+          <p className={classNames('text-dark', styles.info)}>
+            Dive in and learn something new!
+          </p>
           <p className="text-14 text-light">
             {blogList.map((el) => (
               <p className="text-14" key={el.name}>
                 <a
-                  className="link"
+                  className={classNames('link', styles.link)}
                   href={`${el.link}/?ref=shubham.sh`}
                   target="_blank"
                   rel="noreferrer"
