+++ hooks/useIsMounted.ts
@@ -1,0 +1,13 @@
+import * as React from 'react';
+
+function useIsMounted() {
+  const [isMounted, setIsMounted] = React.useState(false);
+
+  React.useEffect(() => {
+    setIsMounted(true);
+  }, []);
+
+  return isMounted;
+}
+
+export default useIsMounted;
