+++ scenes/Home/Home.module.css
@@ -21,30 +21,6 @@
   padding: 0 16px;
 }
 
-.btnMail {
-  border: 1px solid hsla(0, 0%, 88.6%, 0.84);
-  border-radius: 100px;
-  background-color: #fff;
-  box-shadow: 0 1px 4px 0 var(--btn-bsdow);
-  padding-left: 24px;
-  padding-right: 24px;
-  position: relative;
-}
-
-.btnMail:focus {
-  outline: 0;
-}
-
-.btnMail:focus:after {
-  content: '';
-  position: absolute;
-  height: calc(100% + 16px);
-  width: calc(100% + 16px);
-  border-radius: 100px;
-  border: 2px dashed var(--color-purple);
-  top: -8px;
-  left: -8px;
-}
 
 @media (min-width: 1184px) {
   .main {
