+++ styles/theme/light.css
@@ -1,12 +1,16 @@
 :root {
   --bg: #ffffff;
   --color-link-hover: #248aff;
-  --text-color-dark: #2f363f;
+  --text-color-dark: #000;
   --text-color-light: #81878c;
-  --text-color-normal: #000;
+  --text-color-normal: #2f363f;
   --bg-dots: #fafafa;
   --theme-toggler: #805de1;
   --bg-btn: #fff;
   --btn-border: hsla(0, 0%, 88.6%, 0.84);
-  --btn-bsdow: #e2e2e2
+  --btn-bsdow: #e2e2e2;
+  --color-link: #1f7ae0;
+  --color-hand: #c58c85;
+  --color-email: #805de1;
+  --color-link-focus: #805de1;
 }
