+++ styles/utilities.css
@@ -11,11 +11,11 @@
 }
 
 .pointer {
-    cursor: pointer;
+  cursor: pointer;
 }
 
 .relative {
-    position: relative;
+  position: relative;
 }
 
 .inline-block {
@@ -24,4 +24,23 @@
 
 .br-16 {
   border-radius: 16px;
+}
+
+.custom-focus {
+  position: relative;
+}
+
+.custom-focus:focus {
+  outline: 0;
+}
+
+.custom-focus:focus:after {
+  content: '';
+  position: absolute;
+  height: calc(100% + 16px);
+  width: calc(100% + 16px);
+  border-radius: 100px;
+  border: 2px dashed var(--color-link-focus);
+  top: -8px;
+  left: -8px;
 }
